Move hashing code to separate module
And add tests for it.
